change count type long -> Integer

+++ src/main/java/exastro/Exastro_Days_Tokyo/participant_resource/controller/api/v1/ParticipantController.java
@@ -40,10 +40,10 @@ public class ParticipantController extends BaseParticipantController{
 	//セミナー参加人数確認
 	@GetMapping(value = "/count",
 			produces = MediaType.APPLICATION_JSON_VALUE )
-	public long countParticipant(@RequestParam ("seminar_id") int seminarId) {
+	public Integer countParticipant(@RequestParam ("seminar_id") int seminarId) {
 		
 		logger.debug("method called. [ " + Thread.currentThread().getStackTrace()[1].getMethodName() + " ]");
-		Long count = null;	
+		Integer count = null;	
 		try{
 			//セミナー参加人数を取得しリターン
 			count = service.countParticipant(seminarId);
